fix rewrite of runner
