+++ panama/cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli import cli
+
+__all__ = ("cli",)
